Update Program_5.java
Changes made :- 
1) The existing code is calling the sqrt method of Math class again and again via the for loop , which is "computationally expensive" (less optimized) . 
Better option is , to store the square root of the number in another integer variable .

2) The Math.sqrt() method results in giving back a double value , so it is needed to be type casted to an integer variable in the code . Before also , inside the for loop , it was not type casted into an integer which could have lead to various faults/errors.

3) As 1 will be definitely be the factor , so initially sum can be set to the value of 1 , and for loop can be initialized from i = 2 .

+++ Assignment4/Program_5.java
@@ -7,7 +7,7 @@ public class Program_5 {
         System.out.print("Enter a number: ");
         int n = sc.nextInt();
 
-        int sum = 1 , sq_root = Math.sqrt(n);
+        int sum = 1 , sq_root = (int)(Math.sqrt(n));
 
         for (int i = 2; i <= sq_root; i++) {
             if (n % i == 0) {
